[Utils cleanup] Move parameter conversion to the utils module
