+++ test/implementation/implementation_autograd.py
@@ -3,9 +3,8 @@ from .implementation import Implementation
 
 from backpack.hessianfree.hvp import hessian_vector_product
 from backpack.hessianfree.ggnvp import ggn_vector_product_from_plist
-from backpack.hessianfree.lop import L_op
 from backpack.hessianfree.rop import R_op
-from backpack.hessianfree.utils import vector_to_parameter_list
+from backpack.utils.convert_parameters import vector_to_parameter_list
 
 
 class AutogradImpl(Implementation):
